backend/requirements.txt
fastapi[standard]
SQLAlchemy
python-dotenv
python-jose[cryptography]
psycopg2-binary
cryptography
bcrypt
passlib[bcrypt]
supabase==2.22.4
supabase-auth==2.22.4
supabase-functions==2.22.4
httpx
alembic
pandas
python-multipart
joblib
scikit-learn
fpdf2
openpyxl
numpy
xgboost
